refactor: moves configure_test_logger to testutils
The test_logger setup should not be packaged with the
trestlebot code
